+++ test/fixture/src/ES7Decorators.js
@@ -1,5 +1,16 @@
-@decorator
+/**
+ * this is ES7Decorators with decorators
+ */
+@decorator1
+@decorator2
 export class ES7Decorators {
+  /**
+   * this is method with decorators
+   */
+  @decorator1
+  @decorator2
+  method(){}
 }
 
-function decorator(target){}
+function decorator1(target){}
+function decorator2(target){}