+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+- **Fix**
+  - Missing documentation with decorator ([#1](https://github.com/esdoc/esdoc-es7-plugin/issues/1))
+
 ## 0.0.1 (2015-08-09)
 - **Feat**
   - First Release